Historically arp(8) did a route lookup for the entry it is about to add, and failed if it exist and had invalid data link type.
Later on, in r201282, this check morphed to other code, but
message "proxy entry exists for non 802 device" still left,
and now it is printed in a case if route prefix found is
equal to current address being added. In other words, when
we are trying to add ARP entry for a network address. The
message is absolutely unrelated and disappointing in this
case.

I don't see anything bad with setting ARP entries for
network addresses. While useless in usual network,
in a /31 RFC3021 it may be necessary. This, remove this code.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@233773 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f